+++ src/options/RecipeList.vue
@@ -70,15 +70,15 @@
 
 <script>
 import { ref, computed } from "vue";
-import InputText from "@/components/InputText.vue";
-import InputSelect from "@/components/InputSelect.vue";
+import InputText from "@/options/InputText.vue";
+import InputSelect from "@/options/InputSelect.vue";
 import iconTrash from "@/components/icons/iconTrash.vue";
 import iconUpload from "@/components/icons/iconUpload.vue";
 import {
   checkRecipe,
   checkRecipeJson,
   setRecipe,
-} from "@/composables/setRecipe.js";
+} from "@/options/setRecipe.js";
 export default {
   components: { InputText, InputSelect, iconTrash, iconUpload },
   setup() {
